+++ CapitalizeDecorator.java
@@ -1,4 +1,3 @@
-
 public class CapitalizeDecorator extends Decorator {
     public CapitalizeDecorator(Nameable nameable) {
         super(nameable);
