+++ Person.java
@@ -1,16 +1,19 @@
-
+//person class
+import java.util.ArrayList;
+import java.util.List;
 public class Person implements Nameable {
-     int id;
-     String name="unknown";
-     int age;
-     boolean parentPermission=true;
+    int id;
+    String name;
+    int age;
+    boolean parentPermission;
+    List<Rental> rentals;
 
     public Person(String name, int age) {
         this.name = name;
         this.age = age;
-        this.parentPermission = parentPermission;
+        this.parentPermission = true;
+        this.rentals = new ArrayList<>();
     }
-
 
     public int getId() {
         return id;
@@ -43,5 +46,13 @@ public class Person implements Nameable {
     @Override
     public String correctName() {
         return name;
+    }
+
+    public List<Rental> getRentals() {
+        return rentals;
+    }
+
+    public void addRental(Rental rental) {
+        rentals.add(rental);
     }
 }
