--- project Three/CapitalizeDecorator.java
@@ -1,11 +1,0 @@
-public class CapitalizeDecorator extends Decorator {
-    public CapitalizeDecorator(Nameable nameable) {
-        super(nameable);
-    }
-
-    @Override
-    public String correctName() {
-        String name = nameable.correctName();
-        return name.toUpperCase();
-    }
-}
