--- project Three/Decorator.java
@@ -1,15 +1,0 @@
-public abstract class Decorator implements Nameable {
-    protected Nameable nameable;
-
-    public Decorator(Nameable nameable) {
-
-        this.nameable = nameable;
-    }
-
-    @Override
-    public String correctName() {
-
-            return nameable.correctName();
-        }
-    }
-
