--- project Three/Nameable.java
@@ -1,3 +1,0 @@
-public interface Nameable {
-    String correctName();
-}
